finalizado ver1 usando sql puro

+++ app_sql/controllers/controller_tarefas.py
@@ -2,6 +2,8 @@ from models.tarefas import Tarefa
 from repositories.repositorio_tarefas import RepositorioTarefa
 from views.tela_tarefas import TelaTarefas
 
+from datetime import datetime
+
 class ControllerTarefas:
 
     def __init__(self, usuario: int):
@@ -11,7 +13,6 @@ class ControllerTarefas:
     
 
     def adicionar_tarefa(self) -> None:
-
         descricao = input("Descrição: ").strip()
         if descricao == "":
             print("Descrição não pode ficar em branco.\n")
@@ -29,19 +30,20 @@ class ControllerTarefas:
             print("Não foi possível cadastar a tarefa.\n")            
 
 
-    def editar_tarefa(self) -> None:
+    def editar_tarefa(self) -> None:        
+        if not self.existe_tarefas_andamento():
+            return
+
+        self.motrar_tarefas_andamento()
 
         id_tarefa = self.pegar_id_tarefa()
         if id_tarefa == 0: # id inválida
             return self.editar_tarefa()
-        elif id_tarefa == -1: # Não tem tarefas para realizar operação
-            return
         
-        # TODO: arrumar select para pegar apenas campos necessarios
         tarefa_selecionada = self.repo_tarefa.selecionar_tarefa_por_id(id_tarefa)
 
-        print("Informações atuais: ")
-        print(f"\nDescrição: {tarefa_selecionada[2]} - Importância: {tarefa_selecionada[3]}\n") 
+        print("\nInformações atuais: ")
+        print(f"Descrição: {tarefa_selecionada[1]} - Importância: {tarefa_selecionada[2]}\n") 
 
         descricao = input("Nova descrição: (Aperte Enter para não modificar): ")  
 
@@ -50,10 +52,10 @@ class ControllerTarefas:
             importancia = input("Importância: (A = Alta, M = Média, B = Baixa): (Aperte Enter para não modificar) ").upper()
 
         if descricao == "": 
-            descricao = tarefa_selecionada[2]
+            descricao = tarefa_selecionada[1]
         
         if importancia == "":
-            importancia = tarefa_selecionada[3]
+            importancia = tarefa_selecionada[2]
 
         tarefa = Tarefa(self.id_usuario, descricao, importancia, id_tarefa)
 
@@ -65,12 +67,14 @@ class ControllerTarefas:
 
 
     def excluir_tarefa(self) -> None:
+        if not self.existe_tarefas_andamento():
+            return
+
+        self.motrar_tarefas_andamento()
 
         id_tarefa = self.pegar_id_tarefa()
         if id_tarefa == 0: # id inválida
             return self.excluir_tarefa()
-        elif id_tarefa == -1: # Não tem tarefas para realizar operação
-            return
 
         deletado = self.repo_tarefa.deletar_tarefa(id_tarefa)
         if deletado:
@@ -80,12 +84,14 @@ class ControllerTarefas:
         
 
     def finalizar_tarefa(self) -> None:
+        if not self.existe_tarefas_andamento():
+            return
+
+        self.motrar_tarefas_andamento()
 
         id_tarefa = self.pegar_id_tarefa()
         if id_tarefa == 0: # id inválida
             return self.finalizar_tarefa()
-        elif id_tarefa == -1: # Não tem tarefas para realizar operação
-            return
         
         finalizada = self.repo_tarefa.finalizar_tarefa(id_tarefa)
         if finalizada:
@@ -94,64 +100,85 @@ class ControllerTarefas:
             print("Não foi possível finalizar a tarefa.\n")            
 
 
-    def pegar_id_tarefa(self) -> int:
-        # TODO: arrumar select para pegar apenas campos necessarios
+    def existe_tarefas_andamento(self) -> bool:
         tarefas = self.repo_tarefa.selecionar_tarefas_por_status(self.id_usuario, "A")
         if not tarefas:
             print("Usuario não tem tarefas para realizar a operação.\n")
-            return -1
-        
+            return False
+        return True
+
+
+    def motrar_tarefas_andamento(self) -> None:       
         opcao = ""
         while opcao not in ("S", "N"):
             opcao = input("Deseja ver as tarefas em andamento? S/N ").strip().upper()
 
-        # TODO: FORMATAR A EXIBIÇÃO DAS INFORMAÇÕES, USAR VIEWS PARA EXIBIR 
-        if opcao == "S":                        
+        if opcao == "S":
+            tarefas = self.repo_tarefa.selecionar_tarefas_por_status(self.id_usuario, "A")                        
             for tarefa in tarefas:
-                print(tarefa)
+                print(f"Id: {tarefa[0]} - Descrição: {tarefa[1]} - Importância: {tarefa[2]}")
 
+
+    def pegar_id_tarefa(self) -> int:
         id_tarefa = input("Qual o id da tarefa? ")
-        # TODO: arrumar select para pegar apenas campos necessarios
         tarefa_selecionada = self.repo_tarefa.selecionar_tarefa_por_id(id_tarefa)
 
         if not tarefa_selecionada:
             print("Tarefa não cadastrada.\n")
             return 0
 
-        if tarefa_selecionada[1] != self.id_usuario:
+        if tarefa_selecionada[0] != self.id_usuario:
             print("Tarefa não pertence ao usuário.\n")
             return 0
         
-        if tarefa_selecionada[4] != "A":
+        if tarefa_selecionada[3] != "A":
             print("Essa tarefa já foi finalizada.\n")
             return 0
         
         return id_tarefa
 
-    # TODO: arrumar select para pegar apenas campos necessarios
+
     def listar_todas_tarefas(self) -> None:
-        resultado = self.repo_tarefa.selecionar_todas_tarefas()
-        self.tela_tarefas.exibir_tarefas(resultado)
+        tarefas = self.repo_tarefa.selecionar_todas_tarefas(self.id_usuario)
+        if not tarefas:
+            print("Nenhuma tarefa encontrada.\n")
+            return
+        self.tela_tarefas.exibir_tarefas(tarefas)
 
-    # TODO: arrumar select para pegar apenas campos necessarios
-    def listar_tarefas_em_andamento(self) -> None:
-        resultado = self.repo_tarefa.selecionar_tarefas_por_status(self.id_usuario, "A")
-        self.tela_tarefas.exibir_tarefas(resultado)
 
-    # TODO: arrumar select para pegar apenas campos necessarios
-    def listar_tarefas_finalizadas(self) -> None:
-        resultado = self.repo_tarefa.selecionar_tarefas_por_status(self.id_usuario, "F")
-        self.tela_tarefas.exibir_tarefas(resultado)
+    def listar_tarefas_por_status(self, tipo_status: str) -> None:
+        tarefas = self.repo_tarefa.selecionar_tarefas_por_status(self.id_usuario, tipo_status)
+        if not tarefas:
+            print("Nenhuma tarefa encontrada.\n")
+            return
+        self.tela_tarefas.exibir_tarefas(tarefas)
 
-    # TODO: arrumar select para pegar apenas campos necessarios
-    def listar_tarefas_por_importancia(self, importancia: str) -> None:
-        resultado = self.repo_tarefa.selecionar_tarefas_por_importancia(self.id_usuario, importancia)
-        self.tela_tarefas.exibir_tarefas(resultado)
 
-    # TODO: fazer
-    def listar_tarefas_por_criacao(self) -> None:
-        print(f"Listada tarefas por data de criação do usuário {self.id_usuario}")
+    def listar_tarefas_por_importancia(self) -> None:        
+        importancia = None
+        while importancia not in ("A", "M", "B"):
+            importancia = input("Importância: (A = Alta, M = Média, B = Baixa): ").upper()
+            
+        tarefas = self.repo_tarefa.selecionar_tarefas_por_importancia(self.id_usuario, importancia)
+        if not tarefas:
+            print("Nenhuma tarefa encontrada.\n")
+            return
+        self.tela_tarefas.exibir_tarefas(tarefas)
 
-    # TODO: fazer
-    def listar_tarefas_por_finalizacao(self) -> None:
-        print(f"Listada tarefas por data de finalização do usuário {self.id_usuario}")
+
+    def listar_tarefas_por_data(self, tipo_data: str) -> None:
+        data_inicial = input("Data Inicial (dd/mm/aaaa): ").strip()
+        data_final = input("Data Final (dd/mm/aaaa): ").strip()
+
+        try:
+            datetime.strptime(data_inicial, "%d/%m/%Y")
+            datetime.strptime(data_final, "%d/%m/%Y")
+        except ValueError:
+            print("Data no formato errado. Tente novamente.\n")
+            return self.listar_tarefas_por_data(tipo_data)
+
+        tarefas = self.repo_tarefa.selecionar_tarefas_por_data(tipo_data, self.id_usuario, data_inicial, data_final)
+        if not tarefas:
+            print("Nenhuma tarefa encontrada.\n")
+            return
+        self.tela_tarefas.exibir_tarefas(tarefas)
